test(core-backend): move GCS-dependent tests to integration suite
