+++ full-stack-tests/backend/src/integration/GeoServices.test.ts
@@ -8,27 +8,19 @@ import {
   GeographicCRSInterpretRequestProps, GeographicCRSProps,
   Helmert2DWithZOffset,
 } from "@itwin/core-common";
-import { IModelNative } from "../../internal/NativePlatform";
+import { getAvailableCoordinateReferenceSystems, getAvailableCRSUnits, IModelNative } from "@itwin/core-backend";
 import { Geometry, Point3d, Range2d, Range2dProps } from "@itwin/core-geometry";
-import { GeoCoordConfig } from "../../GeoCoordConfig";
-import { getAvailableCoordinateReferenceSystems, getAvailableCRSUnits } from "../../GeographicCRSServices";
-import { TestUtils } from "../TestUtils";
+import { GeoCoordConfig } from "@itwin/core-backend/lib/cjs/GeoCoordConfig";
+import "./StartupShutdown"; // calls startup/shutdown IModelHost before/after all tests
 
 // spell-checker: disable
 
 describe("GeoServices", () => {
   before(async () => {
-    // These tests exercise GCS data loaded from cloud workspaces, so restart the backend with
-    // GCS workspaces enabled (the test harness disables them by default to avoid network calls).
-    await TestUtils.shutdownBackend();
-    await TestUtils.startBackend({ disableGcsWorkspaces: false });
+    // These tests exercise GCS data loaded from cloud workspaces, so ensure the default GCS
+    // databases are loaded. They run as integration tests because they make network requests
+    // to download GCS workspaces from cloud containers.
     GeoCoordConfig.loadDefaultDatabases();
-  });
-
-  after(async () => {
-    // Restore the default test backend so subsequent test suites aren't left with GCS enabled.
-    await TestUtils.shutdownBackend();
-    await TestUtils.startBackend();
   });
 
   const completionTest = async (incompleteGCS: GeographicCRSProps, completeCRS: GeographicCRSProps) => {
